# source/mobile-beta-foundation/services/api/api/openapi.yaml
openapi: 3.1.0
info:
  title: Athletica AI API
  version: 0.5.0
  description: |
    Phase 1 surface of the Athletica AI backend: liveness, authentication with
    logout and token revocation, the idempotent workout-set write, the workout
    lifecycle, the workout history behind the "Итоги" screen and the aggregates
    behind the "Прогресс" screen.

    Three rules hold across the whole contract:

    * the caller's user ID always comes from the bearer access token, never from
      a request body, a query parameter or a cursor;
    * a workout that belongs to another user answers exactly like a workout that
      does not exist (`404`), so existence never leaks;
    * a state change the lifecycle forbids answers `409 invalid_transition` —
      never `500`, and never a silent no-op.

    ### Changes in 0.5.0 (all additive, 0.4.0 clients keep working)

    * `PATCH /workouts/{workoutId}/sets/{setId}` corrects a logged set;
    * `DELETE /workouts/{workoutId}/sets/{setId}` removes one, softly;
    * `PATCH /workouts/{workoutId}` names a workout after it was started —
      the session begun offline with no title is what it exists for;
    * `WorkoutSet` gained the optional `updatedAt` and `deletedAt` fields;
    * error codes `set_deleted` and `workout_not_editable` were added;
    * metrics are exported in Prometheus text form on a **separate** listener
      (`ATHLETICA_METRICS_ADDR`, loopback by default, `ATHLETICA_METRICS_TOKEN`
...

...

...

...

...

...

...
    * `GET /progress`;
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /workouts:
    post:
      tags: [workouts]
      summary: Start a workout owned by the authenticated user, idempotently
      description: |
        The client may name the workout it is starting by sending `id`. That is
        what makes an offline session possible: the phone can begin training
        with no connection and still have one identity to hang its sets on —
        the same reasoning that gives every set a `clientMutationId`.

        Replaying a create with an `id` the caller already holds returns the
        stored workout with `200`, never a second session, and never rewrites
        its title. An `id` held by another account answers `409
        workout_id_taken` without confirming that it exists.

        Omit `id` and the server generates one, exactly as before.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                  description: Client-chosen identity for this workout. Must be a UUID.
                title: { type: string, maxLength: 200 }
      responses:
        '201':
          description: Workout created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workout' }
        '200':
          description: |
            This caller had already started a workout with that `id`; the stored
            row is returned unchanged so a retry after a lost response settles.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workout' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409':
          description: The `id` is already in use by another account (`workout_id_taken`).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422': { $ref: '#/components/responses/ValidationFailed' }

    get:
      tags: [workouts]
      summary: List the caller's workouts, newest first
      description: |
        The history behind the "Итоги" screen. Ordered by `createdAt` descending
        with the workout ID as tiebreaker, and paged with an opaque keyset
        cursor: pass the `nextCursor` of a page back as `cursor` to continue.

        A cursor is opaque and carries no identity — it only ever moves a
        position inside the caller's own rows, so one copied from another
        account cannot widen the result.
      parameters:
        - name: status
          in: query
          required: false
          description: |
            Repeat the parameter or pass a comma-separated list. Omit for every
            status.
          schema:
            type: array
            items: { $ref: '#/components/schemas/WorkoutStatus' }
          style: form
          explode: true
        - name: from
          in: query
          required: false
          description: Inclusive lower bound on `createdAt` (RFC 3339 or `YYYY-MM-DD`, read as UTC).
          schema: { type: string }
        - name: to
          in: query
          required: false
          description: Exclusive upper bound on `createdAt`.
          schema: { type: string }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: cursor
          in: query
          required: false
          description: Opaque `nextCursor` from a previous page.
          schema: { type: string }
      responses:
        '200':
          description: One page of the caller's workouts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkoutPage' }
        '400':
          description: |
            A filter value is malformed — an unknown status, an unparsable date,
            a limit out of range, or a cursor this API did not issue
            (`invalid_cursor`).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /workouts/{workoutId}/sets:
    parameters:
      - name: workoutId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [workouts]
      summary: Log one strength-training set idempotently
      description: |
        The set is stored under the user ID carried by the access token; a
        `userId` field in the body is ignored.

        Idempotency rests on a unique index on `(user_id, client_mutation_id)`
        in PostgreSQL, so a retried outbox entry can never create a second row,
        even when two requests race. Replaying a mutation ID returns `409` with
        the originally stored set — including when the replayed payload differs.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkoutSetInput' }
      responses:
        '201':
          description: Set logged
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkoutSet' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/WorkoutNotFound' }
        '409':
          description: |
            Duplicate client mutation. The body carries the set that was stored
            when this clientMutationId was first accepted.
          content:
            application/json:
              schema:
                type: object
                required: [error, set]
                properties:
                  error: { $ref: '#/components/schemas/ErrorPayload' }
                  set: { $ref: '#/components/schemas/WorkoutSet' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    get:
      tags: [workouts]
      summary: List the sets of one of the caller's workouts
      responses:
        '200':
          description: Sets of the workout, ordered by set number
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/WorkoutSet' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/WorkoutNotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived HS256 access token issued by /auth/login.
  responses:
    BadRequest:
      description: The body is not a single well-formed JSON object
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing, malformed or expired access token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ValidationFailed:
      description: The payload violates the shared domain rules
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    WorkoutNotFound:
      description: |
        The workout does not exist, or belongs to another user. The two cases
        are indistinguishable by design.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    WorkoutStatus:
      type: string
      enum: [active, paused, completed, cancelled]
      description: |
        `active` and `paused` are open states; `completed` and `cancelled` are
        terminal and allow no further transition.
    Workout:
      type: object
      required: [id, title, status, createdAt, updatedAt, endedAt]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        status: { $ref: '#/components/schemas/WorkoutStatus' }
        createdAt: { type: string, format: date-time }
        updatedAt:
          type: string
          format: date-time
          description: When the status last changed. Added in 0.3.0.
        endedAt:
          type: [string, 'null']
          format: date-time
          description: |
            When the workout reached a terminal status; `null` while it is still
            active or paused. Added in 0.3.0.
    WorkoutPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Workout' }
        nextCursor:
          type: [string, 'null']
          description: |
            Opaque position to pass back as `cursor` for the next page; `null`
            on the last page. Clients must echo it untouched and never build one.
...
    WorkoutSetInput:
      type: object
      required: [exerciseId, setNumber, weightKg, repetitions, rir, clientMutationId]
      description: |
        Bounds mirror validateSet() in packages/domain/src/workout.ts. Any
        additional property — `userId` in particular — is ignored by the server.
      properties:
        exerciseId: { type: string, minLength: 1, maxLength: 128 }
        setNumber: { type: integer, minimum: 1 }
        weightKg: { type: number, minimum: 0, maximum: 1000 }
        repetitions: { type: integer, minimum: 1, maximum: 100 }
        rir: { type: integer, minimum: 0, maximum: 10 }
        clientMutationId:
          type: string
          minLength: 1
          maxLength: 128
          description: Stable per-mutation ID from the client outbox.
...
    WorkoutSet:
      type: object
      required: [id, workoutId, exerciseId, setNumber, weightKg, repetitions, rir, clientMutationId, createdAt, updatedAt, deletedAt]
      properties:
        id: { type: string, format: uuid }
        workoutId: { type: string, format: uuid }
        exerciseId: { type: string }
        setNumber:
          type: integer
          description: |
            The number the client gave the set, never reassigned by the server.
            Gaps are normal after a deletion.
        weightKg: { type: number }
        repetitions: { type: integer }
        rir: { type: integer }
        clientMutationId:
          type: string
          description: |
            The ID of the mutation that *created* the set. A correction does not
            replace it.
        createdAt: { type: string, format: date-time }
        updatedAt:
          type: string
          format: date-time
          description: |
            When the set was last corrected; equal to `createdAt` for a set that
            was never edited. Added in 0.5.0.
        deletedAt:
          type: [string, 'null']
          format: date-time
          description: |
            When the athlete removed the set; `null` for every set a client can
            still see. It is non-null only in the answer to a deletion and in
            the `409` a replayed creation of a removed set gets. Added in 0.5.0.
...
    ErrorPayload:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - invalid_request
            - validation_failed
            - unauthorized
            - invalid_credentials
            - email_taken
            - not_found
            - duplicate_client_mutation
            - invalid_transition
            - invalid_cursor
            - workout_id_taken
            - set_deleted
            - workout_not_editable
            - rate_limited
            - internal_error
        message: { type: string }
        details:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
    Error:
      type: object
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ErrorPayload' }